+++ src/ui/handlers/battle-message-ui-handler.ts
@@ -1,16 +1,16 @@
 import { globalScene } from "#app/global-scene";
-import { addBBCodeTextObject, addTextObject, getBBCodeFragment } from "#app/ui/text/text-utils";
-import { TextStyle } from "#enums/text-style";
-import { UiMode } from "#enums/ui-mode";
-import MessageUiHandler from "./message-ui-handler";
-import { addWindow } from "../ui-theme";
-import type BBCodeText from "phaser3-rex-plugins/plugins/bbcodetext";
-import { Button } from "#enums/buttons";
-import i18next from "i18next";
-import type { Stat } from "#enums/stat";
-import { PERMANENT_STATS, getStatKey } from "#enums/stat";
 import { settings } from "#app/system/settings/settings-manager";
 import { GAME_WIDTH, TEXT_SCALE } from "#app/ui-constants";
+import { addBBCodeTextObject, addTextObject, getBBCodeFragment } from "#app/ui/text/text-utils";
+import { addWindow } from "#app/ui/ui-theme";
+import { Button } from "#enums/buttons";
+import type { Stat } from "#enums/stat";
+import { PERMANENT_STATS, getStatKey } from "#enums/stat";
+import { TextStyle } from "#enums/text-style";
+import { UiMode } from "#enums/ui-mode";
+import i18next from "i18next";
+import type BBCodeText from "phaser3-rex-plugins/plugins/bbcodetext";
+import MessageUiHandler from "./message-ui-handler";
 
 export default class BattleMessageUiHandler extends MessageUiHandler {
   private readonly wordWrapWidth: number = (GAME_WIDTH - 24) * TEXT_SCALE;
