+++ src/ui/handlers/egg-summary-ui-handler.ts
@@ -1,18 +1,18 @@
+import { getEggTierForSpecies } from "#app/data/egg";
 import type { EggHatchData } from "#app/data/egg-hatch-data";
 import { globalScene } from "#app/global-scene";
 import { EggSummaryPhase } from "#app/phases/egg-summary-phase";
 import { settings } from "#app/system/settings/settings-manager";
 import { GAME_HEIGHT } from "#app/ui-constants";
+import { HatchedPokemonContainer } from "#app/ui/components/hatched-pokemon-container";
+import PokemonHatchInfoContainer from "#app/ui/components/pokemon-hatch-info-container";
 import { ScrollBar } from "#app/ui/components/scroll-bar";
+import PokemonIconAnimHelper from "#app/ui/helpers/pokemon-icon-anim-helper";
+import ScrollableGridController from "#app/ui/helpers/scrollable-grid-controller";
 import { Button } from "#enums/buttons";
 import { EggSkipPreference } from "#enums/egg-skip-preference";
 import { PokemonIconAnimMode } from "#enums/pokemon-icon-anim-mode";
 import { UiMode } from "#enums/ui-mode";
-import { getEggTierForSpecies } from "#app/data/egg";
-import { HatchedPokemonContainer } from "#app/ui/components/hatched-pokemon-container";
-import PokemonHatchInfoContainer from "#app/ui/components/pokemon-hatch-info-container";
-import PokemonIconAnimHelper from "#app/ui/helpers/pokemon-icon-anim-helper";
-import ScrollableGridController from "#app/ui/helpers/scrollable-grid-controller";
 import MessageUiHandler from "./message-ui-handler";
 
 const iconContainerX = 112;
