+++ src/ui/handlers/loading-modal-ui-handler.ts
@@ -1,8 +1,8 @@
-import i18next from "i18next";
-import { ModalUiHandler } from "./modal-ui-handler";
 import { addTextObject } from "#app/ui/text/text-utils";
 import { TextStyle } from "#enums/text-style";
 import type { UiMode } from "#enums/ui-mode";
+import i18next from "i18next";
+import { ModalUiHandler } from "./modal-ui-handler";
 
 export default class LoadingModalUiHandler extends ModalUiHandler {
   constructor(mode: UiMode | null = null) {
