+++ src/ui/handlers/option-select-ui-handler.ts
@@ -1,5 +1,5 @@
-import type { OptionSelectItem, OptionSelectModeConfig } from "#app/ui/interfaces/option-select-config";
 import AbstractOptionSelectUiHandler from "#app/ui/handlers/abstract-option-select-ui-handler";
+import type { OptionSelectItem, OptionSelectModeConfig } from "#app/ui/interfaces/option-select-config";
 
 export default class OptionSelectUiHandler extends AbstractOptionSelectUiHandler<OptionSelectItem> {
   override show(config?: OptionSelectModeConfig, ..._args: unknown[]): boolean {
