+++ src/ui/handlers/run-history-ui-handler.ts
@@ -1,22 +1,22 @@
 import { globalScene } from "#app/global-scene";
-import { GameModes } from "#enums/game-modes";
-import { addTextObject } from "#app/ui/text/text-utils";
-import { TextStyle } from "#enums/text-style";
-import { UiMode } from "#enums/ui-mode";
-import { addWindow } from "#app/ui/ui-theme";
-import { fixedNumber, getPokemonLevelText, isNullOrUndefined } from "#app/utils";
-import type PokemonData from "#app/system/pokemon-data";
-import MessageUiHandler from "#app/ui/handlers/message-ui-handler";
-import i18next from "i18next";
-import { Button } from "#enums/buttons";
-import { BattleType } from "#enums/battle-type";
 import type { RunEntry } from "#app/system/game-data";
-import { PlayerGender } from "#enums/player-gender";
-import { TrainerVariant } from "#enums/trainer-variant";
-import { RunDisplayMode } from "#enums/run-display-mode";
+import type PokemonData from "#app/system/pokemon-data";
 import { settings } from "#app/system/settings/settings-manager";
 import { GAME_HEIGHT, GAME_WIDTH } from "#app/ui-constants";
+import { addTextObject } from "#app/ui/text/text-utils";
+import { addWindow } from "#app/ui/ui-theme";
+import { fixedNumber, getPokemonLevelText, isNullOrUndefined } from "#app/utils";
+import { BattleType } from "#enums/battle-type";
+import { Button } from "#enums/buttons";
+import { GameModes } from "#enums/game-modes";
 import { ImagesFolder } from "#enums/images-folders";
+import { PlayerGender } from "#enums/player-gender";
+import { RunDisplayMode } from "#enums/run-display-mode";
+import { TextStyle } from "#enums/text-style";
+import { TrainerVariant } from "#enums/trainer-variant";
+import { UiMode } from "#enums/ui-mode";
+import i18next from "i18next";
+import MessageUiHandler from "./message-ui-handler";
 import type RunInfoUiHandler from "./run-info-ui-handler";
 
 type RunSelectCallback = (cursor: number) => void;
