+++ src/ui/handlers/run-info-ui-handler.ts
@@ -1,15 +1,19 @@
 import type { SessionSaveData } from "#app/@types/SessionData";
 import { PLAYER_PARTY_MAX_SIZE } from "#app/constants";
 import { getBiomeName } from "#app/data/balance/biomes";
+import { getNatureName, getNatureStatMultiplier } from "#app/data/nature";
 import { getPokeballAtlasKey } from "#app/data/pokeball";
 import { getTypeRgb } from "#app/data/type";
 import { getVariantTint } from "#app/data/variant";
 import { globalScene } from "#app/global-scene";
+import * as Modifier from "#app/modifier/modifier";
+import { getLuckString, getLuckTextTint } from "#app/modifier/modifier-type";
 import type PokemonData from "#app/system/pokemon-data";
 import { settings } from "#app/system/settings/settings-manager";
 import { DEFAULT_LANGUAGE_KEY } from "#app/system/settings/supported-languages";
 import { GAME_HEIGHT, GAME_WIDTH, TEXT_SCALE } from "#app/ui-constants";
 import { addBBCodeTextObject, addTextObject, getBBCodeFragment } from "#app/ui/text/text-utils";
+import { addWindow } from "#app/ui/ui-theme";
 import {
   formatLargeNumberFixedDigits,
   formatMoney,
@@ -34,11 +38,7 @@ import { TrainerVariant } from "#enums/trainer-variant";
 import { UiMode } from "#enums/ui-mode";
 import i18next from "i18next";
 import RoundRectangle from "phaser3-rex-plugins/plugins/roundrectangle";
-import { getNatureName, getNatureStatMultiplier } from "#app/data/nature";
-import * as Modifier from "#app/modifier/modifier";
-import { getLuckString, getLuckTextTint } from "#app/modifier/modifier-type";
-import { addWindow } from "#app/ui/ui-theme";
-import type UiHandler from "./abstract-ui-handler";
+import UiHandler from "./abstract-ui-handler";
 
 /**
  * RunInfoUiMode indicates possible overlays of RunInfoUiHandler.
