+++ src/ui/handlers/save-slot-select-ui-handler.ts
@@ -1,20 +1,20 @@
 import type { SessionSaveData } from "#app/@types/SessionData";
-import { globalScene } from "#app/global-scene";
-import { fixedNumber, getPlayTimeString, getPokemonLevelText, isNullOrUndefined } from "#app/utils";
-import { Button } from "#enums/buttons";
-import i18next from "i18next";
 import { GameMode } from "#app/game-mode";
+import { globalScene } from "#app/global-scene";
 import * as Modifier from "#app/modifier/modifier";
 import type PokemonData from "#app/system/pokemon-data";
+import { GAME_HEIGHT, GAME_WIDTH } from "#app/ui-constants";
 import type { ConfirmModeConfig } from "#app/ui/interfaces/confirm-menu-config";
-import type MessageUiHandler from "./message-ui-handler";
 import { addTextObject } from "#app/ui/text/text-utils";
+import { addWindow } from "#app/ui/ui-theme";
+import { fixedNumber, getPlayTimeString, getPokemonLevelText, isNullOrUndefined } from "#app/utils";
+import { Button } from "#enums/buttons";
+import { RunDisplayMode } from "#enums/run-display-mode";
+import { SaveSlotUiMode } from "#enums/save-slot-ui-mode";
 import { TextStyle } from "#enums/text-style";
 import { UiMode } from "#enums/ui-mode";
-import { addWindow } from "#app/ui/ui-theme";
-import { SaveSlotUiMode } from "#enums/save-slot-ui-mode";
-import { RunDisplayMode } from "#enums/run-display-mode";
-import { GAME_HEIGHT, GAME_WIDTH } from "#app/ui-constants";
+import i18next from "i18next";
+import MessageUiHandler from "./message-ui-handler";
 import type RunInfoUiHandler from "./run-info-ui-handler";
 
 const SESSION_SLOTS_COUNT = 5;
