+++ src/ui/handlers/target-select-ui-handler.ts
@@ -1,15 +1,15 @@
-import { BattlerIndex } from "#enums/battler-index";
-import { UiMode } from "#enums/ui-mode";
-import UiHandler from "./abstract-ui-handler";
-import { isNullOrUndefined, fixedNumber } from "#app/utils";
 import { getMoveTargets } from "#app/data/moves/move";
+import type { Pokemon } from "#app/field/pokemon";
+import { globalScene } from "#app/global-scene";
+import type { ModifierBar } from "#app/modifier/modifier";
+import { fixedNumber, isNullOrUndefined } from "#app/utils";
 import { isFieldTargeted } from "#app/utils/move-utils";
+import { BattlerIndex } from "#enums/battler-index";
+import { BattlerTagType } from "#enums/battler-tag-type";
 import { Button } from "#enums/buttons";
 import type { MoveId } from "#enums/move-id";
-import type { Pokemon } from "#app/field/pokemon";
-import type { ModifierBar } from "#app/modifier/modifier";
-import { globalScene } from "#app/global-scene";
-import { BattlerTagType } from "#enums/battler-tag-type";
+import { UiMode } from "#enums/ui-mode";
+import UiHandler from "./abstract-ui-handler";
 
 export type TargetSelectCallback = (targets: BattlerIndex[]) => void;
