+++ src/ui/handlers/test-dialogue-ui-handler.ts
@@ -1,10 +1,9 @@
-import type { InputFieldConfig } from "./form-modal-ui-handler";
-import { FormModalUiHandler } from "./form-modal-ui-handler";
-import type { ModalConfig } from "./modal-ui-handler";
-import i18next from "i18next";
+import type { InputFieldConfig, ModalConfig } from "#app/ui/interfaces/modal-config";
 import type { OptionSelectItem, OptionSelectModeConfig } from "#app/ui/interfaces/option-select-config";
 import { isNullOrUndefined } from "#app/utils";
 import { UiMode } from "#enums/ui-mode";
+import i18next from "i18next";
+import { FormModalUiHandler } from "./form-modal-ui-handler";
 
 export default class TestDialogueUiHandler extends FormModalUiHandler {
   keys: string[];
@@ -46,19 +45,19 @@ export default class TestDialogueUiHandler extends FormModalUiHandler {
     this.keys = keys;
   }
 
-  getModalTitle(_config?: ModalConfig): string {
+  getModalTitle(): string {
     return "Test Dialogue";
   }
 
-  getWidth(_config?: ModalConfig): number {
+  getWidth(): number {
     return 300;
   }
 
-  getMargin(_config?: ModalConfig): [number, number, number, number] {
+  getMargin(): [number, number, number, number] {
     return [0, 0, 48, 0];
   }
 
-  getButtonLabels(_config?: ModalConfig): string[] {
+  getButtonLabels(): string[] {
     return ["Check", "Cancel"];
   }
